separate stuff that should be available to other mods into api mod
