+++ api/src/MoreTabs.cpp
@@ -1,4 +1,4 @@
-#include "MoreTabs.hpp"
+#include <MoreTabs.hpp>
 #include <Geode/binding/CCMenuItemToggler.hpp>
 #include <Geode/binding/EditButtonBar.hpp>
 #include <Geode/binding/GameManager.hpp>
@@ -7,6 +7,7 @@
 #include <geode.custom-keybinds/include/Keybinds.hpp>
 
 using namespace keybinds;
+using namespace editor_api;
 
 bool MoreTabs::init(EditorUI* ui) {
     if (!CCNode::init())
